Move color sensor handling to a new file.
By moving these functions to a new file, we keep all color handling code in one place, and allow the main robot code to be cleaner.  The functionality of this commit is basically the same as the previous one: get the color sensor value, check for a match, and send the results to the SmartDashboard.