Initial implementation of Pixy2 class.
Moved the initial code out of robot.py and into pixy2.py.  getSync(), sendPacket(), receivePacket(), and getVersion() are all partially implemented.  We can get the version string from the Pixy, along with searching for sync bytes.
Still to go: implementing the checksum, implementing the Link and SPILink classes, and creating the Version class.